feature: doctor-schedule api
