+++ oupharmacyapp/public/locales/vi/doctor-schedule.json
@@ -1,0 +1,11 @@
+{
+    "monday": "Thứ 2",
+    "tuesday": "Thứ 3",
+    "wednesday": "Thứ 4",
+    "thursday": "Thứ 5",
+    "friday": "Thứ 6",
+    "saturday": "Thứ 7",
+    "sunday": "Chủ nhật",
+    "morning": "Buổi sáng",
+    "afternoon": "Buổi chiều"
+}
