+++ oupharmacyapp/src/modules/pages/DoctorScheduleComponents/services/index.js
@@ -1,0 +1,6 @@
+import APIs, { endpoints } from "../../../../config/APIs"
+
+export const fetchCreateDoctorScheduleByWeek = async (data) => {
+    const res = await APIs.post(endpoints['doctor-schedule-weekly'], data)
+    return res
+}
