+++ oupharmacyapp/src/modules/pages/ExaminationListComponents/hooks/useExaminationList.js
@@ -8,7 +8,6 @@ import { goToTop } from "../../../../lib/utils/helper"
 import UserContext from "../../../../lib/context/UserContext"
 const useExaminationList = () => {
     const {user} = useContext(UserContext)
-    // const [user] = useContext(userContext);
     const [isLoading, setIsLoading] = useState(true)
     const [flag, setFlag] = useState(false)
     const [examinationList, setExaminationList] = useState([])
